# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/checkout/subscribe": {
            "post": {
                "description": "Creates a new payment/subscription",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Billing"
                ],
                "summary": "Subscribe to plan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Plan value in cents",
                        "name": "plan_value",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Billing type (PIX, BOLETO, CREDIT_CARD)",
                        "name": "billing_type",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/billing.PaymentResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/webhook/asaas": {
            "post": {
                "description": "Processes payment webhooks from Asaas",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Billing"
                ],
                "summary": "Process Asaas webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Webhook signature",
                        "name": "asaas-hmac-sha256",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "billing.PaymentResult": {
            "type": "object",
            "properties": {
                "invoiceUrl": {
                    "type": "string"
                },
                "paymentId": {
                    "type": "string"
                },
                "pixCopyPaste": {
                    "type": "string"
                },
                "pixQrCode": {
                    "type": "string"
                }
            }
        },
...
    }
}